requirements.txt
flask-swagger-ui==3.36.0
flask-restx==0.5.0  # Adjusted version
openai==0.27.0
gunicorn==20.1.0
SQLAlchemy==2.0.25
alembic==1.12.1
Werkzeug==3.0.1
pytest==8.0.2
bcrypt==4.1.2
python-dotenv==1.0.1
black==23.1.0 
fastapi>=0.100.0,<0.101.0     
uvicorn[standard]>=0.22.0,<0.23.0
httpx>=0.25.0,<0.26.0
PyPDF2>=3.0.0,<3.1.0